requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Google APIs (for Sheets, Slides, etc.)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Web scraping and automation
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# AI/LLM integration
openai>=1.3.0
groq>=0.4.0
